Fixed Errors

+++ FinalProject/FinalProject/Controllers/Admin/WishlistController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Service.DTOs.Admin.Wishlists;
+using Service.Helpers.Exceptions;
 using Service.Services.Interfaces;
 
 namespace FinalProject.Controllers.Admin
@@ -17,16 +18,15 @@ namespace FinalProject.Controllers.Admin
         public async Task<IActionResult> GetWishlistByUserId(string userId)
         {
             var wishlist = await _wishlistService.GetWishlistByUserIdAsync(userId);
-            if (wishlist == null)
-            {
-                return NotFound();
-            }
+            if (wishlist == null) throw new NotFoundException("User not found");
             return Ok(wishlist);
         }
 
         [HttpPost]
         public async Task<IActionResult> AddWishlist([FromQuery] WishlistDto wishlistDto)
         {
+            var wishlist = await _wishlistService.GetWishlistByUserIdAsync(wishlistDto.ToString());
+            if (wishlist == null) throw new NotFoundException("User not found");
             await _wishlistService.AddWishlistAsync(wishlistDto);
             return CreatedAtAction(nameof(GetWishlistByUserId), new { userId = wishlistDto.AppUserId }, wishlistDto);
         }
@@ -34,8 +34,9 @@ namespace FinalProject.Controllers.Admin
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteWishlist(int id)
         {
+            if (id == null) throw new NotFoundException("Not Found");
             await _wishlistService.DeleteWishlistAsync(id);
-            return NoContent();
+            return Ok();
         }
         [HttpDelete("DeleteProductFromWishlist/{id}")]
         public async Task<IActionResult> DeleteProductFromWishlist(int id, int productId)
